+++ tests/unit/test_init.py
@@ -1,5 +1,7 @@
 """Tests for generic parser."""
 import os
+from pathlib import Path
+import email
 
 import pytest
 
@@ -7,15 +9,17 @@ from circuit_maintenance_parser import (
     init_provider,
     get_provider_class,
     get_provider_class_from_sender,
-    get_provider_data_types,
+    init_data_raw,
+    init_data_email,
+    init_data_emailmessage,
 )
-from circuit_maintenance_parser.errors import NonexistentParserError
-from circuit_maintenance_parser.providers import (
+from circuit_maintenance_parser.data import NotificationData
+from circuit_maintenance_parser.errors import NonexistentProviderError
+from circuit_maintenance_parser.provider import (
     GenericProvider,
     EUNetworks,
     NTT,
     PacketFabric,
-    Telstra,
     Zayo,
 )
 
@@ -23,24 +27,53 @@ from circuit_maintenance_parser.providers import (
 dir_path = os.path.dirname(os.path.realpath(__file__))
 
 
+def test_init_data_raw():
+    """Test the init_data_raw function."""
+    data = init_data_raw("my_type", b"my_content")
+    assert isinstance(data, NotificationData)
+    assert len(data.data_parts) == 1
+    assert data.data_parts[0].type == "my_type"
+    assert data.data_parts[0].content == b"my_content"
+
+
+def test_init_data_email():
+    """Test the email data load."""
+    with open(Path(dir_path, "data", "email", "test_sample_message.eml"), "rb") as email_file:
+        email_raw_data = email_file.read()
+    data = init_data_email(email_raw_data)
+    assert isinstance(data, NotificationData)
+    assert len(data.data_parts) == 7
+
+
+def test_init_data_emailmessage():
+    """Test the emailmessage data load."""
+    with open(Path(dir_path, "data", "email", "test_sample_message.eml"), "rb") as email_file:
+        email_raw_data = email_file.read()
+    raw_email_string = email_raw_data.decode("utf-8")
+    email_message = email.message_from_string(raw_email_string)
+    data = init_data_emailmessage(email_message)
+    assert isinstance(data, NotificationData)
+    assert len(data.data_parts) == 7
+
+
 @pytest.mark.parametrize(
-    "raw, provider_type, result_type",
+    "provider_type, result_type",
     [
-        (b"raw_bytes", "wrong", None),
-        (b"raw_bytes", "", GenericProvider),
-        (b"raw_bytes", "ntt", NTT),
-        (b"raw_bytes", "packetfabric", PacketFabric),
-        (b"raw_bytes", "eunetworks", EUNetworks),
-        (b"raw_bytes", "zayo", Zayo),
+        ("wrong", None),
+        ("", GenericProvider),
+        ("ntt", NTT),
+        ("packetfabric", PacketFabric),
+        ("eunetworks", EUNetworks),
+        ("zayo", Zayo),
     ],
 )
-def test_init_provider(raw, provider_type, result_type):
+def test_init_provider(provider_type, result_type):
     """Tests for init_provider."""
-    result = init_provider(raw=raw, provider_type=provider_type)
+    provider = init_provider(provider_type=provider_type)
     if result_type:
-        assert isinstance(result, result_type)
+        assert isinstance(provider, result_type)
     else:
-        assert result is None
+        assert provider is None
 
 
 @pytest.mark.parametrize(
@@ -54,7 +87,7 @@ def test_init_provider(raw, provider_type, result_type):
         ("Zayo", Zayo, None),
         ("euNetworks", EUNetworks, None),
         ("EUNetworks", EUNetworks, None),
-        ("wrong", None, NonexistentParserError),
+        ("wrong", None, NonexistentProviderError),
     ],
 )
 def test_get_provider_class(provider_name, result, error):
@@ -62,7 +95,7 @@ def test_get_provider_class(provider_name, result, error):
     if result:
         assert get_provider_class(provider_name) == result
     elif error:
-        with pytest.raises(NonexistentParserError):
+        with pytest.raises(error):
             get_provider_class(provider_name)
 
 
@@ -73,7 +106,7 @@ def test_get_provider_class(provider_name, result, error):
         (NTT.get_default_organizer(), NTT, None),
         (Zayo.get_default_organizer(), Zayo, None),
         (EUNetworks.get_default_organizer(), EUNetworks, None),
-        ("wrong", None, NonexistentParserError),
+        ("wrong", None, NonexistentProviderError),
     ],
 )
 def test_get_provider_class_from_email(email_sender, result, error):
@@ -81,30 +114,5 @@ def test_get_provider_class_from_email(email_sender, result, error):
     if result:
         assert get_provider_class_from_sender(email_sender) == result
     elif error:
-        with pytest.raises(NonexistentParserError):
+        with pytest.raises(error):
             get_provider_class_from_sender(email_sender)
-
-
-@pytest.mark.parametrize(
-    "provider_name, result, error",
-    [
-        ("packetfabric", PacketFabric.get_data_types(), None),
-        ("pAcketfabrIc", PacketFabric.get_data_types(), None),
-        ("ntt", NTT.get_data_types(), None),
-        ("NTT", NTT.get_data_types(), None),
-        ("zayo", Zayo.get_data_types(), None),
-        ("Zayo", Zayo.get_data_types(), None),
-        ("euNetworks", EUNetworks.get_data_types(), None),
-        ("EUNetworks", EUNetworks.get_data_types(), None),
-        ("TelstRa", Telstra.get_data_types(), None),
-        ("Telstra", Telstra.get_data_types(), None),
-        ("wrong", None, NonexistentParserError),
-    ],
-)
-def test_get_provider_data_types(provider_name, result, error):
-    """Tests for to validate the data types for each provider"""
-    if result:
-        assert get_provider_data_types(provider_name) == result
-    elif error:
-        with pytest.raises(NonexistentParserError):
-            get_provider_data_types(provider_name)